feat(pace-93): fetch motd message from getstrm.com

+++ cmd/pace/main.go
@@ -6,6 +6,8 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/cobra/doc"
 	"github.com/spf13/pflag"
+	"io"
+	"net/http"
 	"os"
 	"pace/pace/pkg/bootstrap"
 	"pace/pace/pkg/common"
@@ -43,7 +45,7 @@ var RootCmd = &cobra.Command{
 
 func rootCmdPreRun(cmd *cobra.Command, _ []string) error {
 	CreateConfigDirAndFileIfNotExists()
-	SetLastSeenTimestamp()
+	SetLastSeenTimestamp(cmd)
 	err := bootstrap.InitializeConfig(cmd)
 	log.Infoln(fmt.Sprintf("Executing command: %v", cmd.CommandPath()))
 	cmd.Flags().Visit(func(flag *pflag.Flag) {
@@ -81,20 +83,22 @@ func CreateConfigDirAndFileIfNotExists() {
 	}
 }
 
-func SetLastSeenTimestamp() {
-	lastSeenCommandFilepath := path.Join(common.ConfigPath(), common.DefaultLastSeenFilename)
-	now := time.Now()
-	if content, err := os.ReadFile(lastSeenCommandFilepath); err != nil {
-		updateLastSeen(lastSeenCommandFilepath, now)
-	} else {
-		ts, err := strconv.ParseInt(strings.Trim(string(content), "\n"), 10, 64)
-		if err != nil {
-			os.Remove(lastSeenCommandFilepath)
+func SetLastSeenTimestamp(cmd *cobra.Command) {
+	if cmd.Name() != "__complete" {
+		lastSeenCommandFilepath := path.Join(common.ConfigPath(), common.DefaultLastSeenFilename)
+		now := time.Now()
+		if content, err := os.ReadFile(lastSeenCommandFilepath); err != nil {
 			updateLastSeen(lastSeenCommandFilepath, now)
-		}
-		lastSeen := time.Unix(ts, 0)
-		if lastSeen.Unix() < now.Add(-24*time.Hour).Unix() {
-			updateLastSeen(lastSeenCommandFilepath, now)
+		} else {
+			ts, err := strconv.ParseInt(strings.Trim(string(content), "\n"), 10, 64)
+			if err != nil {
+				os.Remove(lastSeenCommandFilepath)
+				updateLastSeen(lastSeenCommandFilepath, now)
+			}
+			lastSeen := time.Unix(ts, 0)
+			if lastSeen.Unix() < now.Add(-24*time.Hour).Unix() {
+				updateLastSeen(lastSeenCommandFilepath, now)
+			}
 		}
 	}
 }
@@ -109,17 +113,13 @@ func updateLastSeen(lastSeenCommandFilepath string, now time.Time) {
 }
 
 func printWelcomeMessage() {
-	asciiArt := `
----------------------------------------------------------------------
-                        ____   _    ____ _____                         
-                       |  _ \ / \  / ___| ____|
-                       | |_) / _ \| |   |  _|  
-                       |  __/ ___ \ |___| |___ 
-                       |_| /_/   \_\____|_____|
----------------------------------------------------------------------
-Hey there, cool you're using PACE!
-We'd love to learn what your use case is or if you have any feedback.
-Join our Slack: https://getstrm.com/slack
----------------------------------------------------------------------`
-	fmt.Println(asciiArt)
+	res, _ := http.Get("https://cli.getstrm.com/motd")
+	if res.StatusCode != 200 {
+		log.Warnln("Could not fetch MOTD message, non-200 response")
+		return
+	}
+
+	resBody, _ := io.ReadAll(res.Body)
+
+	fmt.Println(string(resBody))
 }
